Refactor AddCommand
Now AddCommand is abstract, and the original functionality is moved to
AddPersonCommand

Update AddCommandParser to handle different cases of `add` command.
